Move handlers to ui layer
Move entire handlers directory to src/ui/handlers as part of architecture refactoring
to consolidate all UI-related code.
