+++ src/ui/handlers/image_loading.py
